feat(go): Dynamic Get/Set Property Calls
Added runtime code gen for calls to jsii get/set apis. Changed order of
args in runtime Get/Set functions to better match method invoke.

Moved all property access to codegen to separate file.
