+++ packages/jsii-pacmak/lib/targets/go/runtime/constants.ts
@@ -23,7 +23,7 @@ export const JSII_GET_FUNC = `${JSII_RT_ALIAS}.Get`;
 export const JSII_SGET_FUNC = `${JSII_RT_ALIAS}.StaticGet`;
 
 // JSII set request
-export const JSII_SET_FUNC = `${JSII_RT_ALIAS}.Get`;
+export const JSII_SET_FUNC = `${JSII_RT_ALIAS}.Set`;
 
 // JSII static set request
 export const JSII_SSET_FUNC = `${JSII_RT_ALIAS}.StaticSet`;
@@ -35,3 +35,10 @@ export const JSII_OVERRIDE = `${JSII_RT_ALIAS}.Override`;
 export const JSII_ANY = `${JSII_RT_ALIAS}.Any`;
 // Jsii FQN type
 export const JSII_FQN = `${JSII_RT_ALIAS}.FQN`;
+
+// NOOP type returns
+export const NOOP_RETURN_MAP: { [type: string]: string } = {
+  float64: '0.0',
+  string: '"NOOP_RETURN_STRING"',
+  bool: 'true',
+};
