+++ packages/jsii-pacmak/lib/targets/go/runtime/index.ts
@@ -1,4 +1,5 @@
 export * from './constants';
 export * from './class-constructor';
 export * from './method-call';
+export * from './property-access';
 export * from './util';
